plug-ins: fix criticals in sketchbook tiff causing cairo assert
While testing with the poc from issue #16251 criticals were shown
in the terminal:
GLib-GObject-CRITICAL **: 19:10:44.016: value "-2147483648.000000" of
 type 'gdouble' is invalid or out of range for property 'y' of
 type 'gdouble'
Eventually this crashed GIMP with an assert inside cairo.

Due to image_height being unsigned and an unusually large layer, the
partial result underflowed and thus resulted in a very large value,
which caused criticals when calling gimp_layer_set_offsets.
That eventually causes issues and an assert in cairo, most likely
when trying to create a preview for the layer.

We fix this by casting the unsigned values to gfloat, since that is
also what the final result is expected to be.

While we are at it, also promote the 32-bit row value a few lines
before this to gsize to make sure we always compute the correct offset.

(cherry picked from commit 61eeee891664138b9d72ef059f2e16f155a80d1e)

+++ plug-ins/file-tiff/file-tiff-load.c
@@ -2644,7 +2644,7 @@ load_sketchbook_layers (TIFF      *tif,
                                        GEGL_RECTANGLE (0, layer_height - row - 1,
                                                        layer_width, 1),
                                        0, format,
-                                       ((guchar *) pixels) + row * layer_width * 4,
+                                       ((guchar *) pixels) + (gsize) row * layer_width * 4,
                                        GEGL_AUTO_ROWSTRIDE);
                     }
                   g_object_unref (buffer);
@@ -2653,9 +2653,9 @@ load_sketchbook_layers (TIFF      *tif,
 
               if (! GIMP_IS_GROUP_LAYER (layer))
                 {
-                  x_pos += (layer_width -
+                  x_pos += ((gfloat) layer_width -
                             gimp_drawable_get_width (GIMP_DRAWABLE (layer)));
-                  y_pos = image_height -
+                  y_pos = (gfloat) image_height -
                           gimp_drawable_get_height (GIMP_DRAWABLE (layer)) - y_pos;
 
                   gimp_layer_set_offsets (layer, ROUND (x_pos), ROUND (y_pos));
